Checkpoint: add modport parsing, misc parse fixes

+++ 201102_parser_updates1/plugins/net.sf.sveditor.core/src/net/sf/sveditor/core/db/index/cache/SVDBDirFS.java
@@ -56,7 +56,9 @@ public class SVDBDirFS implements ISVDBFS {
 	
 	public void delete(String path) {
 		if (path.equals("")) {
-			delete_tree(fDBDir);
+			if (fDBDir.exists()) {
+				delete_tree(fDBDir);
+			}
 		} else {
 			File file = new File(fDBDir, path);
